Switch int to double

+++ profile.h
@@ -1,8 +1,8 @@
 #ifndef PROFILE_H_
 #define PROFILE_H_
 
-constexpr int kMaxVelocity = 1.0;
-constexpr int kMaxAcceleration = 1.0;
+constexpr double kMaxVelocity = 1.0;
+constexpr double kMaxAcceleration = 1.0;
 
 class Profile {
 public:
